refactor(radio-button): prevent consumption of RadioButtonMapper component
Addresses FE-5017

BREAKING CHANGE: `RadioButtonMapper` is now an internal component and can no longer be
imported by consuming apps.
